Refactor naming validator into modular naming package
