+++ calculogic-validator/src/naming/naming-validator.contracts.mjs
@@ -1,0 +1,1 @@
+export const NAMING_VALIDATOR_CONTRACTS_VERSION = '0.1.6';
