+++ calculogic-validator/src/naming/naming-validator.host.mjs
@@ -1,0 +1,1 @@
+export * from './naming-validator.wiring.mjs';
