+++ calculogic-validator/src/naming/naming-validator.wiring.mjs
@@ -1,0 +1,1 @@
+export * from './naming-validator.logic.mjs';
