+++ calculogic-validator/src/naming/registries/naming-case-rules.knowledge.mjs
@@ -1,0 +1,1 @@
+export const CANONICAL_SEMANTIC_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
