+++ calculogic-validator/src/naming/registries/naming-extensions.knowledge.mjs
@@ -1,0 +1,11 @@
+export const REPORTABLE_EXTENSIONS = new Set([
+  '.ts',
+  '.tsx',
+  '.js',
+  '.jsx',
+  '.mjs',
+  '.cjs',
+  '.css',
+  '.json',
+  '.md',
+]);
