+++ calculogic-validator/src/naming/rules/naming-rule-check-semantic-case.logic.mjs
@@ -1,0 +1,3 @@
+import { CANONICAL_SEMANTIC_PATTERN } from '../registries/naming-case-rules.knowledge.mjs';
+
+export const isCanonicalSemanticName = semanticName => CANONICAL_SEMANTIC_PATTERN.test(semanticName);
